dentalapp-backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple/
asgiref==3.2.10
cffi==1.14.3
cryptography==3.3.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django==3.1.2
djangorestframework==3.12.1
phonenumbers==8.12.11
psycopg2==2.8.6
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.4.1